+++ packages/otel/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @ai-sdk/otel
 
+## 1.0.94
+
+### Patch Changes
+
+- Updated dependencies [a4ba394]
+- Updated dependencies [36b3364]
+- Updated dependencies [45099da]
+  - ai@7.0.94
+  - @ai-sdk/provider@4.0.11
+
 ## 1.0.93
 
 ### Patch Changes